pyproject.toml
[tool.poetry]
name = "demo-analytic-app"
version = "0.1.0"
description = "Simple demo analytic app"
authors = ["Denis Novikov <alpden550@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
maya = "^0.6.1"
django-admin-rangefilter = "^0.8.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = "^22.1.0"
mixer = "^7.2.1"
flake8 = "^4.0.1"
pytest-django = "^4.5.2"
flake8-bugbear = "^22.1.11"
pep8-naming = "^0.12.1"
flake8-commas = "^2.1.0"
flake8-bandit = "^2.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
